requirements.txt
mltu==0.1.4
tensorflow<=2.10
Pillow~=9.5.0
stow~=1.2.2
numpy~=1.25.0
flask~=2.3.3
keras~=2.10.0
opencv-python~=4.8.0.76
pytesseract~=0.3.10